requirements.txt
python-telegram-bot==21.6
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
anthropic>=0.39.0
httpx>=0.27.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
alpaca-py>=0.30.0
aiosqlite>=0.20.0
pydantic>=2.8.0
